moved company location info to MapActivity
